+++ crates/cli/tests/reference/targets.rs
@@ -3,7 +3,7 @@
 // FLAGS: --target=no-modules
 // FLAGS: --target=nodejs
 // FLAGS: --target=deno
-// FLAGS: --target=source-phase
+// FLAGS: --target=module
 
 use wasm_bindgen::prelude::*;